requirements.txt
pyyaml
requests
pydantic
openai
python-dotenv
psutil
nltk
tabulate
numpy
colorama
ruamel.yaml
